fixed bug with token

+++ internal/entities/token.go
@@ -8,5 +8,5 @@ type Token struct {
 	UserID       uuid.UUID `db:"user_id"`
 	Username     string    `db:"-"`
 	RefreshToken string    `db:"token"`
-	ExpTime      int64     `db:"expiration_time"`
+	ExpTime      float64   `db:"expiration_time"`
 }
